refactor: move xattr tests to a different file
